Treasure Hunter
A random map is created based on the map size information from the user, and the character's start and end points are determined. An algorithm has been designed to enable the character to collect treasure chests as quickly as possible and reach the target. The shortcut algorithm is used for this purpose.